Changed structure. Removed IDE specific files.
