+++ Interview.FilterJson/pom.xml
@@ -1,12 +1,15 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    
     <modelVersion>4.0.0</modelVersion>
     
-    <groupId>Interview</groupId>
+    <parent>
+        <groupId>Interview</groupId>
+        <artifactId>Interview.Parent</artifactId>
+        <version>1.0</version>
+    </parent>
     <artifactId>Interview.FilterJson</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    
+
     <dependencies>
         <dependency>
             <groupId>org.glassfish</groupId>
@@ -14,5 +17,5 @@
             <version>1.1.2</version>
         </dependency>
     </dependencies>
-    
+
 </project>
