+++ Interview.FilterJson/src/main/java/interview/DataExtractionAttribute.java
