+++ Interview.WrapText/src/main/java/interview/WrapText.java
@@ -1,0 +1,22 @@
+package interview;
+
+/**
+ * Implement a function that takes an arbitrary natural language text and wraps it, such that:
+ * 
+ * <ol>
+ *   <li>Every line is indented with four spaces</li>
+ *   <li>No line is longer than 80 characters in length</li>
+ *   <li>Words are not cut in the middle</li>
+ * </ol>
+ */
+
+public final class WrapText
+{
+    public static String wrap( final String text )
+    {
+        // TODO
+        
+        return null;
+    }
+    
+}
